requirements.txt
python-telegram-bot>=20.0
groq>=0.4.0
python-dotenv>=1.0.0